requirements.txt
langchain-core
langchain-community
langchain-groq
python-dotenv
streamlit